+++ src/extension.ts
@@ -39,6 +39,7 @@ export function activate(context: vscode.ExtensionContext) {
         { f: transformutilities.urlDecode, id: 'editor.urlDecode' },
         { f: transformutilities.regnize, id: 'editor.regnize' },
         { f: transformutilities.headerToBookmark, id: 'editor.headerToBookmark' },
+        { f: transformutilities.mixer, id: 'editor.mixer' },
         { f: lineutilities.removeDuplicatedLines, id: 'editor.removeDuplicatedLines' },
         { f: lineutilities.removeEmptyLines, id: 'editor.removeEmptyLines' },
         { f: lineutilities.joinLines, id: 'editor.joinLines' },