+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - (DEV) Add support for multiple inputbox in `utilitymanager`.
  > This addition allows for tools to request multiple inputs from the user
+- Add command: mixer
 
 
 ## [1.0.2] - 2018-08-22